Add oldest, newest, oldestIndex, and newestIndex support
This is notably lacking nth, which is partially implemented but
throws an unimplemented error since p-nth doesn't return YAML
output and peek doesn't provide a simple nth option.

This comes with corresponding tests, though it should be noted
that the tests are not robust due to the asynchronous nature of
the functions being tested. Converting this to a promise API
would be a cleaner solution for reliable tests.