Fix `process_protein` to only call the callback once
There was a logic error in `process_protein` that was causing the
callback to be called twice. In a previous patch, the function was
augmented to call `callback(null)` when the protein couldn't be parsed
properly. This is reasonable, but `callback(null)` was being executed
even after a successful `callback(protein)`, because of where the
`callback(null)` was (after the `catch` block.