Make tests more deterministic
* empty_pool was being called without waiting for the p-stop and
  p-create processes to finish.  I rewrote it to order those processes
  correctly, return a promise, and fulfill the promise when both are
  done. (nb: no error checking is done for either process)
* Use constants for pool names to avoid confusion.
* The "poke" calls to deposit proteins into a pool weren't guaranteed to
  deposit in the expected order.  Fix this by using a chain of promises
  to serialize the writes to and spawns of the poke processes.

Unfortunately, the oldestIndex and newestIndex tests are broken on my
machine for reasons I don't understand.  The functions work just fine in
a Node REPL, but under the tests it seems like the stdout of
p-oldest-idx/p-newest-idx is closed before any data (i.e. "the index")
is written to them.